GPT_Judgment_System_v12_Clean_Installer/requirements.txt
openai==1.22.0
streamlit
python-dotenv
